:recycle: schemaのファイルを外に出す
