+++ src/index.ts
@@ -1,46 +1,8 @@
 import { ApolloServer } from '@apollo/server';
 import { startStandaloneServer } from '@apollo/server/standalone';
 import { GraphQLScalarType } from 'graphql';
-
-const typeDefs = `
-  enum PhotoCategory {
-    SELFIE
-    PORTRAIT
-    ACTION
-    LANDSCAPE
-    GRAPHIC
-  }
-  type User {
-  githubLogin: ID!
-  name: String
-  avatar: String
-  postedPhotos: [Photo!]!
-  inPhotos: [Photo!]!
-  }
-  scalar DateTime
-  type Photo {
-    id: ID!
-    url: String!
-    name: String!
-    description: String
-    category: PhotoCategory!
-    postedBy: User!
-    taggedUsers: [User!]!
-    created: DateTime!
-  }
-  type Query {
-    totalPhotos: Int!
-    allPhotos(after:DateTime): [Photo!]!
-  }
-  input PostPhotoInput {
-    name: String!
-    category: PhotoCategory=PORTRAIT
-    description: String
-  }
-  type Mutation {
-    postPhoto(input:PostPhotoInput!): Photo!
-  }
-`
+//TODO .jsなくてもよくしたい
+import { typeDefs } from './schema.js'
 
 // 1. ユニークIDをインクリメントするための変数
 let _id = 0
