+++ src/schema.ts
@@ -1,0 +1,40 @@
+//TODO .grapqlファイルにしたかったが上手くimportできなかったので後で
+export const typeDefs = `
+  enum PhotoCategory {
+    SELFIE
+    PORTRAIT
+    ACTION
+    LANDSCAPE
+    GRAPHIC
+  }
+  type User {
+  githubLogin: ID!
+  name: String
+  avatar: String
+  postedPhotos: [Photo!]!
+  inPhotos: [Photo!]!
+  }
+  scalar DateTime
+  type Photo {
+    id: ID!
+    url: String!
+    name: String!
+    description: String
+    category: PhotoCategory!
+    postedBy: User!
+    taggedUsers: [User!]!
+    created: DateTime!
+  }
+  type Query {
+    totalPhotos: Int!
+    allPhotos(after:DateTime): [Photo!]!
+  }
+  input PostPhotoInput {
+    name: String!
+    category: PhotoCategory=PORTRAIT
+    description: String
+  }
+  type Mutation {
+    postPhoto(input:PostPhotoInput!): Photo!
+  }
+`
